Auto-wire shutdown; UGC preview/metadata fixes; overlay fields; stale constants
steam_api_shutdown marked finish_fn in spec.gmidl instead of hand-editing
the generated injection stub - Shutdown_Steamworks() now looks up and
calls it automatically on game end, so persistent callbacks no longer
survive game_restart. Confirmed via GM-ExtensionGenerator's own source
(GmidlSchemaParser.cs/CppInjectorsEmitter.cs) and matched against real
precedent already shipping in GMEXT-Discord/GMEXT-WebView.

steam_ugc_get_query_ugc_additional_preview's original_file_name was
never a real input - the SDK parameter it maps to is STEAM_OUT_STRING_COUNT
(output only), and the value was being probed then thrown away. Removed
the dead input param, added original_file_name to
SteamUgcAdditionalPreview as a real output field, updated the one live
demo call site (obj_steam_ugc/Create_0.gml) to match the new 3-arg
signature.

steam_ugc_get_query_ugc_metadata's hardcoded 1024-byte buffer replaced
with k_cchDeveloperMetadataMax (5000), the SDK's own documented max.

GameOverlayActivated_t's m_bUserInitiated/m_nAppID added to
SteamFriendsGameOverlayActivated - GML could not previously tell a
user-triggered overlay toggle from an automatic one.

Four stale constants corrected against the vendored SDK:
STEAM_INVENTORY_INTERFACE_VERSION V002->V003, STEAM_INPUT_INTERFACE_VERSION
SteamInput001->SteamInput006, STEAM_INPUT_MAX_ANALOG_ACTIONS 16->24,
STEAM_INPUT_MAX_DIGITAL_ACTIONS 128->256.

steam_remote_storage_get_file_timestamp's int64 SDK return was narrowed
to int32 - now returns the real int64.

Build-verified: cmake --preset win-x64-release, exit 0.

+++ source/Steamworks_gml/extensions/steamworks/source/src/gm_steam_remote_storage.cpp
@@ -247,13 +247,13 @@ std::int32_t steam_remote_storage_get_file_size(std::string_view file_name)
     return (std::int32_t)rs->GetFileSize(fn.c_str());
 }
 
-std::int32_t steam_remote_storage_get_file_timestamp(std::string_view file_name)
+std::int64_t steam_remote_storage_get_file_timestamp(std::string_view file_name)
 {
     STEAM_GUARD_RET(0);
     ISteamRemoteStorage* rs = steam_remote_storage_iface();
     if (!rs) return 0;
     std::string fn(file_name);
-    return (std::int32_t)rs->GetFileTimestamp(fn.c_str());
+    return (std::int64_t)rs->GetFileTimestamp(fn.c_str());
 }
 
 std::int32_t steam_remote_storage_get_file_count()
